Setting up the creation routine.
