+++ NamespaceFixer/NamespaceBuilder/CsNamespaceBuilderService.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 using System.Text.RegularExpressions;
 
 namespace NamespaceFixer.NamespaceBuilder
@@ -10,31 +9,14 @@ namespace NamespaceFixer.NamespaceBuilder
         {
         }
 
-        public override bool UpdateFile(ref string fileContent, string desiredNamespace)
+        protected override Match GetNamespaceMatch(string fileContent)
         {
-            var namespaceMatch = Regex.Match(fileContent, "namespace\\s([^\n{]+)");
+            return Regex.Match(fileContent, "namespace\\s([^\n{]+)");
+        }
 
-            if (!namespaceMatch.Success)
-            {
-                return false;
-            }
-
-            var fileRequiresUpdate = false;
-            foreach (var group in namespaceMatch.Groups)
-            {
-                if (group is Match match)
-                {
-                    var currentNamespace = match.Value.Trim().Split(' ').Last().Trim();
-
-                    if (currentNamespace != desiredNamespace)
-                    {
-                        fileRequiresUpdate = true;
-                        fileContent = fileContent.Replace(BuildNamespaceLine(currentNamespace), BuildNamespaceLine(desiredNamespace));
-                    }
-                }
-            }
-
-            return fileRequiresUpdate;
+        protected override string BuildNamespaceLine(string desiredNamespace)
+        {
+            return "namespace " + desiredNamespace;
         }
 
         internal override string BuildNamespaceAccordingToOptions(
@@ -60,11 +42,6 @@ namespace NamespaceFixer.NamespaceBuilder
             replaceWithFormat(NamespaceSections.FileToProjectPath, fileToProjectPath);
 
             return newNamespace;
-        }
-
-        private string BuildNamespaceLine(string desiredNamespace)
-        {
-            return "namespace " + desiredNamespace;
         }
     }
 }
